+++ changelog.md
@@ -5,6 +5,8 @@
 * Horse armor can now be recycled
 * Solarium no longer increases attack speed
 * Increased Gravity Defying and M.A. Jump cost
+* Soul Steel now requires Netherite Scrap to be made
+  * Can also be made in a crafting table like Netherite
 * Like the base mod, materials now have a percentage increase/decrease in attack speed
 * Fixed missing Quaron Nugget recipe
 * Fixed enchanting table screen having no background